Complete potential aesthetic changes
The potential change in aeasthetic seen in the previous commit has been applied to all segments of the page, with the previously implemented alternating L/R remaining. A Twitter feed has also been added, and the structure of the site under the hood has changed, with a 'scripts' folder being added to contain jQuery, the Twitter module, the ScrollTo module (to hopefully be used in future to aid the function of the navbar and tooltip), and of course the external JavaScript.