Alter colour scheme and navbar stylisation
Colours for picpage segments have been altered to achieve a balance between vibrancy (which is something Jo asked for specifically IIRC), suitability for text and other content, and aesthetic (i.e. the difference between 'too boring' and 'in-your-face') - this is kind of subjective though so please alter the colours at your whim. :-)
I also changed the navbar from upper-case to the opposite to allow for consistency. Thoughts?